moved things to the model directory and fixed a test by skipping it
