+++ src/cascade/model/__init__.py
@@ -1,2 +1,7 @@
-from cascade.core.log import getLoggers
-CODELOG, MATHLOG = getLoggers(__name__)
+from cascade.model.dismod_groups import DismodGroups
+from cascade.model.model import Model, model_from_vars
+from cascade.model.session import Session
+from cascade.model.smooth_grid import SmoothGrid
+from cascade.model.var import Var
+
+__all__ = [Model, Session, DismodGroups, SmoothGrid, Var, model_from_vars]
