+++ src/cascade/model/smooth_grid.py
