+++ src/cascade/model/var.py
